[#59] TrackEditor: Improvements and bugfixes (Part 89: Fixed false-positive marking of bits as "badly encoded")

+++ Main/src/Image.h
@@ -457,7 +457,7 @@
 			Codec::TType codec;
 			Medium::TType mediumType;
 			BYTE nConsecutiveZerosMax; // # of consecutive zeroes to lose synchronization; e.g. 3 for MFM code
-			BYTE lastReadBits;
+			WORD lastReadBits; // validity flag and bit, e.g. 10b = valid bit '0', 11b = valid bit '1', 0Xb = invalid bit 'X'
 
 			CTrackReader(PLogTime logTimes,DWORD nLogTimes,PCLogTime indexPulses,BYTE nIndexPulses,Medium::TType mediumType,Codec::TType codec,TDecoderMethod method,bool resetDecoderOnIndex);
